requirements.txt
python-nmap
requests
pycvesearch